requirements.txt
allure-pytest
pytest
pytest-check
pydantic-settings